Move last_updated_date to model.py
